+++ packages/bridging/src/providers/across/const/contracts.ts
@@ -29,6 +29,8 @@ export const ACROSS_SPOKE_POOL_CONTRACT_ADDRESSES: Record<TargetChainId, string 
   // Not supported chains
   [SupportedChainId.GNOSIS_CHAIN]: undefined,
   [SupportedChainId.AVALANCHE]: undefined,
+  [AdditionalTargetChainId.BITCOIN]: undefined,
+  [AdditionalTargetChainId.SOLANA]: undefined,
 }
 
 export const ACROSS_SPOKE_POOL_PERIPHERY_CONTRACT_ADDRESSES: Record<TargetChainId, string | undefined> = {
@@ -47,4 +49,7 @@ export const ACROSS_SPOKE_POOL_PERIPHERY_CONTRACT_ADDRESSES: Record<TargetChainI
   [SupportedChainId.GNOSIS_CHAIN]: undefined,
   [SupportedChainId.SEPOLIA]: undefined,
   [SupportedChainId.AVALANCHE]: undefined,
+  [AdditionalTargetChainId.OPTIMISM]: undefined,
+  [AdditionalTargetChainId.BITCOIN]: undefined,
+  [AdditionalTargetChainId.SOLANA]: undefined,
 }